+++ src/main/java/com/limechain/beefy/state/BeefyState.java
@@ -1,16 +1,9 @@
 package com.limechain.beefy.state;
 
 import com.limechain.ServiceConsensusState;
-import com.limechain.beefy.dto.BeefyPayloadId;
-import com.limechain.beefy.dto.Commitment;
-import com.limechain.beefy.dto.PayloadElement;
 import com.limechain.beefy.dto.SignedCommitment;
 import com.limechain.beefy.dto.ValidatorSet;
 import com.limechain.beefy.dto.VoteMessage;
-import com.limechain.exception.beefy.BeefyGenericException;
-import com.limechain.network.protocol.beefy.messages.consensus.BeefyConsensusMessage;
-import com.limechain.network.protocol.warp.DigestHelper;
-import com.limechain.network.protocol.warp.dto.BlockHeader;
 import com.limechain.runtime.Runtime;
 import com.limechain.state.AbstractState;
 import com.limechain.storage.DBConstants;
@@ -29,7 +22,6 @@ import java.math.BigInteger;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
 
 /**
@@ -49,7 +41,6 @@ public class BeefyState extends AbstractState implements ServiceConsensusState {
     private ValidatorSet validatorSet;
 
     private BigInteger disabledAuthority;
-    private byte[] mmrRootHash;
 
     private final BlockState blockState;
     private final KeyStore keyStore;
@@ -167,7 +158,7 @@ public class BeefyState extends AbstractState implements ServiceConsensusState {
         return repository.find(DBConstants.BEEFY_SET_ID, BigInteger.ZERO);
     }
 
-    public void persistValidatorsSetId() {
+    private void persistValidatorsSetId() {
         repository.save(DBConstants.BEEFY_SET_ID, validatorSet.getSetId());
     }
 
@@ -179,7 +170,7 @@ public class BeefyState extends AbstractState implements ServiceConsensusState {
         );
     }
 
-    public void persistBeefyValidators() {
+    private void persistBeefyValidators() {
         repository.save(
                 StateUtil.generateAuthorityKey(DBConstants.BEEFY_AUTHORITY_SET, validatorSet.getSetId()),
                 validatorSet.getValidators()
@@ -214,23 +205,5 @@ public class BeefyState extends AbstractState implements ServiceConsensusState {
                 StateUtil.generateBeefyJustificationKey(DBConstants.BEEFY_JUSTIFICATION, blockNumber),
                 justification
         );
-    }
-
-    private Commitment getCommitment(BigInteger blockNumber, BigInteger setId) {
-        BlockHeader blockHeader = blockState.getHeaderByNumber(blockNumber);
-        byte[] mmrHash = extractMmrRootHash(blockHeader);
-        PayloadElement payloadElement = new PayloadElement(BeefyPayloadId.MMR, mmrHash);
-
-        return new Commitment(Collections.singletonList(payloadElement), blockNumber, setId);
-    }
-
-    private byte[] extractMmrRootHash(BlockHeader blockHeader) {
-        return DigestHelper.getBeefyConsensusMessages(blockHeader.getDigest())
-                .stream().map(BeefyConsensusMessage::getMmrRootHash)
-                .filter(Objects::nonNull)
-                .findFirst()
-                .orElseThrow(() -> new BeefyGenericException(
-                        String.format("No MMR digest found in block header: %d", blockHeader.getBlockNumber()))
-                );
     }
 }
